refactor: move quota::flavor_quota types from lib to wire crate
